requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.3
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
SQLAlchemy==2.0.49
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2026.2
